+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in CUPS v2.4.2 (TBA)
 - Fixed copyright in CUPS Web UI trailer template (Issue #346)
 - mDNS hostname in device uri is not resolved when installaling a permanent
   IPP Everywhere queue (Issues #340, #343)
+- Re-added LibreSSL/OpenSSL support (Issue #???)
 
 
 Changes in CUPS v2.4.1 (27th January 2020)